// src/basic/enums/WindowOffset.ts
export const enum WindowOffset {
    LEFT_TOP = "left_top",
    LEFT_CENTER = "left_center",
    LEFT_BOTTOM = "left_bottom",
    TOP_CENTER = "top_center",
    CENTER = "center",
    BOTTOM_CENTER = "bottom_center",
    RIGHT_TOP = "right_top",
    RIGHT_CENTER = "right_center",
    RIGHT_BOTTOM = "right_bottom"
}